+++ ALUM/ALUM/Views/ProfileViews/MenteeProfileView 2.swift
@@ -11,7 +11,6 @@ import WrappingHStack
 struct MenteeProfileView: View {
     @StateObject private var viewModel = MenteeProfileViewmodel()
     @State var isAtTop: Bool = true
-    @State var uID: String = ""
     var body: some View {
         GeometryReader { grr in
             VStack(spacing: 0) {
@@ -35,7 +34,7 @@ struct MenteeProfileView: View {
                                 Circle()
                                     .frame(width: 135, height: 145)
                                     .foregroundColor(Color("ALUM White2"))
-                                Image("ALUMLogoBlue")
+                                viewModel.mentee.profilePic
                                     .resizable()
                                     .frame(width: 135, height: 135)
                                     .clipShape(Circle())
@@ -43,19 +42,19 @@ struct MenteeProfileView: View {
                             }
                             .padding(.top, 57)
                         }
-                        Text(viewModel.menteeGET.mentee.name)
+                        Text(viewModel.mentee.name)
                             .font(Font.custom("Metropolis-Regular", size: 34, relativeTo: .largeTitle))
                     }
                     HStack {
                         Image(systemName: "graduationcap")
                             .frame(width: 25.25, height: 11)
                             .foregroundColor(Color("ALUM Primary Purple"))
-                        Text(String(viewModel.menteeGET.mentee.grade) + "th Grade @ NHS")
+                        Text(viewModel.mentee.grade + "th Grade @ NHS")
                             .font(Font.custom("Metropolis-Regular", size: 17, relativeTo: .headline))
                     }
                     .padding(.bottom, 18)
                     if !viewModel.selfView {
-                        WhyPairedComponent(text: viewModel.menteeGET.mentee.whyPaired ?? "")
+                        WhyPairedComponent()
                             .padding(.bottom, 16)
                     }
                     Text("About")
@@ -64,16 +63,16 @@ struct MenteeProfileView: View {
                         .frame(maxWidth: .infinity, alignment: .leading)
                         .padding(.leading, 16)
                         .padding(.bottom, 8)
-                    Text(viewModel.menteeGET.mentee.about )
+                    Text(viewModel.mentee.mentorshipGoal)
                         .font(Font.custom("Metropolis-Regular", size: 17, relativeTo: .headline))
                         .lineSpacing(5)
                         .frame(maxWidth: .infinity, alignment: .leading)
                         .padding(.horizontal, 16)
                         .padding(.bottom, 32)
-                    RenderTags(tags: viewModel.menteeGET.mentee.careerInterests, title: "Career Interests")
+                    RenderTags(tags: viewModel.mentee.careerInterests, title: "Career Interests")
                         .padding(.leading, 16)
                         .padding(.bottom, 8)
-                    RenderTags(tags: viewModel.menteeGET.mentee.topicsOfInterest, title: "Topics of Interest")
+                    RenderTags(tags: viewModel.mentee.topicsOfInterest, title: "Topics of Interest")
                         .padding(.leading, 16)
                         .padding(.bottom, 8)
                     if viewModel.selfView {
@@ -85,7 +84,7 @@ struct MenteeProfileView: View {
                             .padding(.leading, 16)
                             .padding(.bottom, 8)
                         MentorCard(
-                            name: viewModel.menteeGET.mentee.pairingId ?? "Name", major: "CS",
+                            name: viewModel.mentee.mentor, major: "CS",
                         university: "UCSD", career: "Software Engineer",
                         profilePic: Image("TestMenteePFP"), isEmpty: false)
                             .padding(.bottom, 10)
@@ -122,16 +121,6 @@ struct MenteeProfileView: View {
                 }
             }
         }
-        .onAppear(perform: {
-            Task{
-                do{
-                    try await viewModel.getMenteeInfo(userID: uID)
-                }
-                catch{
-                    print("Error")
-                }
-            }
-        })
     }
 }
